# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
...
paths:
...
  /api/skills/clone:
    get:
      operationId: discoverSkills
      summary: Discover skills from repo path
      parameters:
        - name: repoPath
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Discovered skills
    post:
      operationId: cloneSkillRepo
      summary: Clone a skill repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                skillsDir:
                  type: string
      responses:
        "200":
          description: Cloned
...
  /api/clone:
    get:
      operationId: listClonedRepos
      summary: List cloned repositories
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items:
                      type: object
    post:
      operationId: cloneRepo
      summary: Clone a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Cloned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
                  name:
                    type: string
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
                  existed:
                    type: boolean
    patch:
      operationId: switchBranch
      summary: Switch branch on cloned repo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repoPath, branch]
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
  /api/clone/progress:
    post:
      operationId: cloneWithProgress
      summary: Clone with SSE progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
      responses:
        "200":
          description: SSE progress stream
          content:
            text/event-stream: {}
...